separated create pool form logic into custom hook
- solved "double posting" problem by making more rigorous checks on newListing data
- removed unnecessary form variable in other hooks since react-hook-form already holds that state
