+++ swimster-ui/src/components/Register/Register.jsx
@@ -4,7 +4,7 @@ import { useRegistrationForm } from '../../hooks/useRegistrationForm'
 import { style } from './style'
 
 const Register = () => {
-    const { form, error, isProcessing, register, handleSubmit, onSubmit } = useRegistrationForm()
+    const { error, isProcessing, register, handleSubmit, onSubmit } = useRegistrationForm()
 
     return (
         <div className={style.register}>
